# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Books API",
        "description": "API REST pour la gestion de livres avec authentification",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "Livres"
                ],
                "summary": "Liste paginée des livres",
                "description": "Retourne la liste des livres avec pagination (2 livres par page)",
                "operationId": "921c595372ab399821e267fcf05cbaa1",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des livres récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Livres"
                ],
                "summary": "Créer un nouveau livre",
                "description": "Crée un nouveau livre dans la base de données (authentification requise)",
                "operationId": "7ab92284110ace5d0f7de52f372e70c0",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer {token}"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livre créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Livres"
                ],
                "summary": "Afficher un livre spécifique",
                "description": "Retourne les détails d'un livre (avec mise en cache pendant 60 minutes)",
                "operationId": "e079e587c1fee53fe150984526abf3ef",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du livre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livre non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Livres"
                ],
                "summary": "Modifier un livre",
                "description": "Met à jour les informations d'un livre existant (authentification requise)",
                "operationId": "b47b1606570ba86a105bf62f04de6b7f",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer {token}"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livre modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livre non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Livres"
                ],
                "summary": "Supprimer un livre",
                "description": "Supprime un livre de la base de données (authentification requise)",
                "operationId": "95c7fc7d78f9dc3884a5bb898b7ead11",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer {token}"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Livre supprimé avec succès"
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livre non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BookResource": {
                "properties": {
                    "data": {
                        "properties": {
                            "title": {
                                "type": "string",
                                "example": "1984"
                            },
                            "author": {
                                "type": "string",
                                "example": "GEORGE ORWELL"
                            },
                            "summary": {
                                "type": "string",
                                "example": "Roman dystopique..."
                            },
                            "isbn": {
                                "type": "string",
                                "example": "9780451524935"
                            },
                            "_links": {
                                "properties": {
                                    "self": {
                                        "properties": {
                                            "href": {
                                                "type": "string",
                                                "example": "http://localhost:8000/api/v1/books/1"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "update": {
                                        "properties": {
                                            "href": {
                                                "type": "string",
                                                "example": "http://localhost:8000/api/v1/books/1"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "delete": {
                                        "properties": {
                                            "href": {
                                                "type": "string",
                                                "example": "http://localhost:8000/api/v1/books/1"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "all": {
                                        "properties": {
                                            "href": {
                                                "type": "string",
                                                "example": "http://localhost:8000/api/v1/books"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Book": {
                "required": [
                    "title",
                    "author",
                    "summary",
                    "isbn"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 3,
                        "example": "1984"
                    },
                    "author": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 3,
                        "example": "GEORGE ORWELL"
                    },
                    "summary": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 10,
                        "example": "Roman dystopique décrivant une société totalitaire contrôlée par Big Brother."
                    },
                    "isbn": {
                        "type": "string",
                        "maxLength": 13,
                        "minLength": 13,
                        "example": "9780451524935"
                    }
                },
                "type": "object"
            },
...
            "UnauthenticatedError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "NotFoundError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "No query results for model [App\\Models\\Book]."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field": [
                                "The field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}